Extract compare utils to their own file
